Se elimina OrdenInput y se reemplaza con SapOrden

+++ Interface_SAP-QS3/src/main/java/com/danone/bonafont/batch/dao/ArchivoDAO.java
@@ -8,5 +8,5 @@ import com.danone.bonafont.batch.model.Archivo;
  */
 public interface ArchivoDAO extends GenericDAO<Archivo, Long>{
 
-	Integer registerFile(String fileName, Integer status, Integer interfaz);
+	Long registerFile(String fileName, Integer status, Integer interfaz);
 }
